update 386 to new runtime (CL 30381)
R=r
DELTA=298  (119 added, 81 deleted, 98 changed)
OCL=30427
CL=30443

+++ src/pkg/runtime/linux/os.h
@@ -3,8 +3,8 @@
 // license that can be found in the LICENSE file.
 
 // Linux-specific system calls
-int64	futex(uint32*, int32, uint32, Timespec*, uint32*, uint32);
-int64	clone(int32, void*, M*, G*, void(*)(void));
+int32	futex(uint32*, int32, uint32, Timespec*, uint32*, uint32);
+int32	clone(int32, void*, M*, G*, void(*)(void));
 
 struct Sigaction;
 void	rt_sigaction(int64, struct Sigaction*, void*, uint64);
